requirements.txt
python-telegram-bot[job-queue]==22.7
anthropic>=0.96.0
python-dotenv>=1.2.2
Pillow>=10.0.0
aiofiles>=23.0.0
flask>=3.0.0
flask-cors>=4.0.0
openpyxl>=3.1.0
gunicorn>=21.2.0
